requirements.txt
asgiref==3.7.2
Django==5.0.2
factory-boy==3.3.0
Faker==23.2.1
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
